+++ src/operations/init.rs
@@ -9,7 +9,7 @@ use std::path::Path;
 ///
 /// This is used to apply the theme when your shell is opened. It is based on your previously applied
 /// theme or your default theme set in config.
-pub fn init(config_path: &Path, data_path: &Path) -> Result<()> {
+pub fn init(config_path: &Path, data_path: &Path, is_verbose: bool) -> Result<()> {
     let config = Config::read(config_path)?;
     let active_scheme_name = fs::read_to_string(data_path.join(CURRENT_SCHEME_FILE_NAME))
         .unwrap_or(config.default_scheme.clone().unwrap_or_default());
@@ -18,7 +18,7 @@ pub fn init(config_path: &Path, data_path: &Path) -> Result<()> {
         return Err(anyhow!("Failed to initialize, config files seem to be missing. Try applying a theme first with `{} apply <SCHEME_NAME>`.", REPO_NAME));
     }
 
-    operations::apply::apply(config_path, data_path, active_scheme_name.as_str())
+    operations::apply::apply(config_path, data_path, active_scheme_name.as_str(), !is_verbose)
             .with_context(|| {
                 format!(
                     "Failed to initialize {}, config files are missing. Try applying a theme first with `{} apply <SCHEME_NAME>`.",